Struct plc_diagnostics::diagnostician::Diagnostician

pub struct Diagnostician { /* private fields */ }

the Diagnostician handle’s Diangostics with the help of a assessor and a reporter

Implementations

impl Diagnostician

pub fn register_file(&mut self, id: String, src: String) -> usize

registers the given source-code at the diagnostician, so it can preview errors in the source returns the id to use to reference the given file

pub fn handle(&mut self, diagnostics: &[Diagnostic]) -> Severity

Assess and reports the given diagnostics.

pub fn null_diagnostician() -> Diagnostician

Creates a null-diagnostician that does not report diagnostics

pub fn buffered() -> Diagnostician

Creates a buffered-diagnostician that saves its reports in a buffer

pub fn clang_format_diagnostician() -> Diagnostician

Creates a clang-format-diagnostician that reports diagnostics in clang format

pub fn with_configuration(self, configuration: DiagnosticsConfiguration) -> Self

pub fn explain(&self, error: &str) -> String

Explain the error with the given code by consulting the diagnostics registry

pub fn get_diagnostic_configuration(&self) -> String