Struct plc_diagnostics::diagnostics::Diagnostic

pub struct Diagnostic { /* private fields */ }

The Diagnostics struct describes an issue encountered during compile time. The issue is defined by an error_code and had a defined severity Diagnostic severity can be overridden when being reported.

Implementations

impl Diagnostic

Builder for Diagnostics

pub fn new(message: impl Into<String>) -> Self

pub fn with_location<T>(self, location: T) -> Self
where T: Into<SourceLocation>,

pub fn with_secondary_location<T>(self, location: T) -> Self
where T: Into<SourceLocation>,

pub fn with_secondary_locations(self, locations: Vec<SourceLocation>) -> Self

pub fn with_error_code(self, code: &'static str) -> Self

pub fn with_sub_diagnostic(self, diagnostic: Diagnostic) -> Self

pub fn with_sub_diagnostics(self, diagnostics: Vec<Diagnostic>) -> Self

pub fn with_internal_error(self, error: Error) -> Self

pub fn from_serde_error(error: Error, source: &SourceCode) -> Self

impl Diagnostic

pub fn get_message(&self) -> &str

pub fn get_location(&self) -> SourceLocation

pub fn get_secondary_locations(&self) -> Option<&[SourceLocation]>

pub fn get_error_code(&self) -> &'static str

pub fn get_sub_diagnostics(&self) -> &[Diagnostic]

impl Diagnostic

pub fn unexpected_token_found( expected: &str, found: &str, range: SourceLocation ) -> Diagnostic

pub fn missing_function(location: SourceLocation) -> Diagnostic

pub fn codegen_error( message: impl Into<String>, location: SourceLocation ) -> Diagnostic

pub fn llvm_error(file: &str, llvm_error: &str) -> Diagnostic

pub fn missing_token(expected_token: &str, range: SourceLocation) -> Diagnostic

pub fn invalid_argument_count<T>( expected: usize, actual: usize, location: T ) -> Diagnostic
where T: Into<SourceLocation>,

pub fn unknown_type(type_name: &str, location: SourceLocation) -> Diagnostic

pub fn unresolved_reference( reference: &str, location: SourceLocation ) -> Diagnostic

pub fn invalid_assignment( right_type: &str, left_type: &str, location: SourceLocation ) -> Diagnostic

pub fn cannot_generate_initializer( variable_name: &str, location: SourceLocation ) -> Diagnostic

pub fn cannot_generate_call_statement(operator: &AstNode) -> Diagnostic

pub fn cannot_generate_from_empty_literal( type_name: &str, location: SourceLocation ) -> Diagnostic

impl Diagnostic

pub fn unnamed_control(location: SourceLocation) -> Diagnostic