plc_diagnostics::diagnostics

Struct Diagnostic

pub struct Diagnostic {
    pub inner: Box<DiagnosticsInner>,
}

The Diagnostics struct describes an issue encountered during compile time. The issue is defined by an error_code and had a defined severity Diagnostic severity can be overridden when being reported.

Fields
inner: Box<DiagnosticsInner>
Implementations

impl Diagnostic

Builder for Diagnostics

pub fn new(message: impl Into<String>) -> Self

pub fn with_location<T>(self, location: T) -> Self
where T: Into<SourceLocation>,

pub fn with_secondary_location<T>(self, location: T) -> Self
where T: Into<SourceLocation>,

pub fn with_secondary_locations(self, locations: Vec<SourceLocation>) -> Self

pub fn with_error_code(self, code: &'static str) -> Self

pub fn with_sub_diagnostic(self, diagnostic: Diagnostic) -> Self

pub fn with_sub_diagnostics(self, diagnostics: Vec<Diagnostic>) -> Self

pub fn with_internal_error(self, error: Error) -> Self

pub fn from_serde_error(error: Error, source: &SourceCode) -> Self

impl Diagnostic

pub fn get_message(&self) -> &str

pub fn get_location(&self) -> SourceLocation

pub fn get_secondary_locations(&self) -> Option<&[SourceLocation]>

pub fn get_error_code(&self) -> &'static str

pub fn get_sub_diagnostics(&self) -> Vec<&Diagnostic>

impl Diagnostic

pub fn unexpected_token_found<T>( expected: &str, found: &str, location: T, ) -> Diagnostic
where T: Into<SourceLocation>,

pub fn missing_function<T>(location: T) -> Diagnostic
where T: Into<SourceLocation>,

pub fn codegen_error<T, U>(message: T, location: U) -> Diagnostic
where T: Into<String>, U: Into<SourceLocation>,

pub fn llvm_error(file: &str, llvm_error: &str) -> Diagnostic

pub fn missing_token<T>(expected_token: &str, location: T) -> Diagnostic
where T: Into<SourceLocation>,

pub fn invalid_argument_count<T>( expected: usize, actual: usize, location: T, ) -> Diagnostic
where T: Into<SourceLocation>,

pub fn unknown_type<T>(type_name: &str, location: T) -> Diagnostic
where T: Into<SourceLocation>,

pub fn unresolved_reference<T>(reference: &str, location: T) -> Diagnostic
where T: Into<SourceLocation>,

pub fn invalid_assignment<T>( right_type: &str, left_type: &str, location: T, ) -> Diagnostic
where T: Into<SourceLocation>,

pub fn cannot_generate_initializer<T>( variable_name: &str, location: T, ) -> Diagnostic
where T: Into<SourceLocation>,

pub fn cannot_generate_call_statement(operator: &AstNode) -> Diagnostic

pub fn cannot_generate_from_empty_literal<T>( type_name: &str, location: T, ) -> Diagnostic
where T: Into<SourceLocation>,

pub fn const_pragma_is_not_allowed<T>(location: T) -> Diagnostic
where T: Into<SourceLocation>,

impl Diagnostic

pub fn unnamed_control<T>(location: T) -> Diagnostic
where T: Into<SourceLocation>,